create(config): filePatterns
