+++ config/filePatterns.go
@@ -1,0 +1,29 @@
+package config
+
+import (
+	"fmt"
+	"strings"
+
+	"github.com/bmatcuk/doublestar/v4"
+)
+
+type SConfigFilrPatterns struct {
+	Include []string
+	Exclude []string
+}
+
+func (this SConfigFilrPatterns) ValidateValues() error {
+	var invalidPatterns = make([]string, 0)
+
+	for _, pattern := range append(Values.FilePatterns.Include, Values.FilePatterns.Exclude...) {
+		if !doublestar.ValidatePathPattern(pattern) {
+			invalidPatterns = append(invalidPatterns, pattern)
+		}
+	}
+
+	if len(invalidPatterns) > 0 {
+		return fmt.Errorf("The following file pattens are invalid:\n%s", strings.Join(invalidPatterns, "\n"))
+	}
+
+	return nil
+}
